refactoring model renderer. ellipse node view centering
